refactor: postContent에서 로직 분리하기
