+++ src/pages/PostsManagerPage.tsx
@@ -2,7 +2,7 @@ import { Plus } from "lucide-react"
 import { Card, CardHeader, CardTitle, Button } from "@/shared/ui"
 
 import { DialogContainer } from "@/widgets/dialogContainer/ui/DialogContainer"
-import { PostContent } from "@/widgets/ui/PostContent"
+import { PostContent } from "@/widgets/postContent/ui/PostContent"
 
 import { useDialogStore } from "@/features/dialogs/model/useDialogStore"
